pkg/util/k8sutil -> pkg/k8sutil, pkg/tlsutil -> pkg/tls
