+++ pkg/scaffold/cmd.go
@@ -44,7 +44,7 @@ import (
 	"{{ .Repo }}/pkg/apis"
 	"{{ .Repo }}/pkg/controller"
 
-	k8sutil "github.com/operator-framework/operator-sdk/pkg/util/k8sutil"
+	"github.com/operator-framework/operator-sdk/pkg/k8sutil"
 	sdkVersion "github.com/operator-framework/operator-sdk/version"
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
